update after 202606 forecast:
- bugfix in runtime diagnostics:
  diagnostics/anom_CPS1_runtime.sh
  diagnostics/plot_forecast_all_vars.sh

- bugfix in qa_checker_template
  postproc/C3S_standard/launch_c3s_qa_checker.sh
  templates/launch_c3s_qa_checker.template.sh

- reduced number of parallel postproc to avoid issues due to too many jobs on queue
  postproc/C3S_standard/launch_postproc_C3S_forecast.sh

- bugfix in month counting
  util/check_status_forecasts.sh

- sleep added to avoid problem in the submission of postpc_clm (dependency)
  postproc/clm/create_clm_files.sh

- triplette file - forecast 202606
  ../triplette_done/triplette.random.202606.txt

- moved sendmail for recover jobs in order to be sent every time a member is resubmitted
  recover/recover_interrupted.sh

- modification to tackle 2013 in extended forecast
  modified:   scripts_oper/create_branch_extended.sh
  modified:   recover/refresh_all_scripts.sh

+++ src/postproc/nemo/interp_ORCA2_1X1_gridT2C3S_template.sh
@@ -185,14 +185,20 @@ do
              exit 1
           else
              touch ${check_oceregrid}
-             rm $inputfile
+             if [[ -f $inputfile ]]
+             then 
+                rm $inputfile
+             fi
           fi
       else
          sleep 300
       fi
    else
       touch ${check_oceregrid}
-      rm $inputfile
+      if [[ -f $inputfile ]]
+      then
+         rm $inputfile
+      fi
       break
    fi
 done
